Updated resolvers with email authorization from json web token

+++ resolvers/Mutation.js
@@ -6,10 +6,13 @@ const jwt = require('jsonwebtoken');
 // Log in user
 loginUser = async (_, { userName }) => {
   try {
+    // Find user
     const user = await User.findOne({ userName: userName });
 
+    // If user does not exist throw error.
     if (!user) throw new Error('This user has not been registered.');
 
+    // Create token
     const token = await jwt.sign({ id: user._id, email: user.email }, '123456');
 
     return { user, token };
@@ -21,11 +24,14 @@ loginUser = async (_, { userName }) => {
 // Add & register a user
 registerUser = async (_, { userName, email, password }) => {
   try {
+    // Check if user is already registered. If registered throw error.
     let user = await User.findOne({ email: email });
     if (user) throw new Error(`${email} has already been registered`);
 
+    // Create new user
     user = await User.create({ userName, email, password });
 
+    // Create token
     const token = await jwt.sign({ id: user._id, email: user.email }, '123456');
 
     return { user, token };
@@ -35,11 +41,16 @@ registerUser = async (_, { userName, email, password }) => {
 };
 
 // Update a user (email)
-updateUser = async (_, { id, email }) => {
+updateUser = async (_, { id, email }, { email: userEmail }) => {
   try {
+    // Verify user is logged in.
+    if (!userEmail) throw new Error('Not Authenticated.');
+
+    // Find user or throw error
     let user = await User.findOne({ _id: id });
     if (!user) throw new Error('This user does not exist!');
 
+    // Find user and update
     user = await User.findOneAndUpdate({ _id: id }, { email: email });
 
     return user;
@@ -49,11 +60,16 @@ updateUser = async (_, { id, email }) => {
 };
 
 // Delete a user
-deleteUser = async (_, { email }) => {
+deleteUser = async (_, { email }, { email: userEmail }) => {
   try {
+    // Verify user is logged in
+    if (!userEmail) throw new Error('Not Authenticated.');
+
+    // Check if user exists. If no throw error.
     let user = await User.findOne({ email });
     if (!user) throw new Error(`${email} doesn't exist.`);
 
+    // Find user and delete
     user = await User.findOneAndDelete({ email });
 
     return user;
@@ -64,24 +80,35 @@ deleteUser = async (_, { email }) => {
 
 // Add a movie
 addMovie = async (_, { userId, title, genre, releaseDate }, { email }) => {
-  // Find current User
-  const user = await User.findOne({ email: email });
+  try {
+    // Verify user is logged in.
+    if (!email) throw new Error('Not Authenticated.');
 
-  // Create a Movie
-  const movie = await Movie.create({ userId, title, genre, releaseDate });
+    // Find current User
+    const user = await User.findOne({ email: email });
 
-  // Push our movie to the user in movies array.
-  user.movies.push(movie);
+    // Create a Movie
+    const movie = await Movie.create({ userId, title, genre, releaseDate });
 
-  // Save Movie
-  user.save();
+    // Push our movie to the user in movies array.
+    user.movies.push(movie);
 
-  return movie;
+    // Save Movie
+    user.save();
+
+    return movie;
+  } catch (error) {
+    return error;
+  }
 };
 
 // Delete a movie
-deleteMovie = async (_, { id }) => {
+deleteMovie = async (_, { id }, { email }) => {
   try {
+    // Verify user is logged in.
+    if (!email) throw new Error('Not Authenticated.');
+
+    // Find and delete movie
     const movie = await Movie.findOneAndDelete({ _id: id });
 
     return movie;
